scripts/requirements.txt
boto3
requests
PyYAML
gitpython
